Added in ACP Adminer
where the Adminer app is made for standalone use, it is included here with toolbox cause it is technically a dev tool. adds in adminer to the acp for database. this is a kinda "hacky" implementation, but everything appears to work. I am curious if this can some how be used to replace the faulty schema manager from IPS dev center? but that would probably be a bad idea.