fixing DEV_DISABLE_ACP_SESSION_TIMEOUT
this on its own does not work if you have a cron job setup to execute task (like i have configured to test certain features on a few of my apps) as there is a task that runs that clears the session. I have overridden a method in \IPS\Session abstract class to only delete sessions that are over 20 years old (a bit overkill i know :)) but this only works if DEV_DISABLE_ACP_SESSION_TIMEOUT is set to true in the constants.php. also added a task to run every hour if this constant is set, to run the admin notifications (runChecksAndSendNotifications).