added in a sockets debug
i am probably the only one who will ever use this, i like using the debug that is apart of profiler, but with multiple tabs opened, the long polling script can some times cause issues, so the sockets replaces this functionality. it needs some legwork to setup, like enabling it thru a constant (which can be found on the toolbox tab for changed constants) and it also requires  exporting a packages.json and nodejs app from toolbox's settings page, can be found in the devtoolbox menu bar once the DT_NODE has been enabled.