WIP: profiler Console
since IPS throws a lot of junk in the console now days, i'm working on a profile console, that you can add a special JS class to your app, that will send console messages to the profiler console instead of the browser one. right now its a work in progress and not fully implemented, but you can test it out if you like. you can also configure it to replace browsers console to use this class instead.