Beta Uploader Task
This will likely will only ever be used by me, its a task to upload files in my "beta" folder on my dev to my site for my beta testers. you can take a look at the code and see if you can use it or modify it to your needs, but this feature will likely remain only for me.