pushing before martin a. has a snarky comment about commit size :P
added in parsers for query log and IPS log on the profiler, so the back/stack traces get parsed out and linked if using DEV_WHOOPS_EDITOR, and a compatible editor. also some other tweaks, like objects/arrays passed to any of the consoles or any of the other tabs like JS Vars and Environment get formatted to a collapsable tree. fixed class look up for hooks.