various updates
https://ips.test/index.php?app=toolbox&module=bt&controller=build&myApp=formularize&do=queue
https://ips.test/index.php?app=toolbox&module=bt&controller=build&myApp=formularize&download=1&do=queue

code analyzer improvements: added in a source folder checker, this is similar to IPS's scanner.php, but with several improvements (note this is not based on their code). will scan only IPS extended classes looking for various things like return types mismatch, parameter mismatches (this includes, but not limited too parameter types/hints and missing/changed paramaters, including default values or options/required changes), signature mismatching, etc.

adding the following to any docblock of any comment will bypass a few of the checks in the code analyzer:
@ips-lint ignore - will make code analyzer completely ignore this method.
@ips-lint ignore-signature - will not perform the signature check on the method.
@ips-lint ignore-parameters - will not perform the parameter check on the method.
@ips-lint ignore-parent - will not perform the parent call check. this will be useful for methods that are replacing the parent method and don't need to call it. the class scanner will have a few of these it will ignore automatically.

this differs from the IPS scanner.php in a few ways, first i use reflectionclass and AST lib to do my checks, so it should understand things like imports and not require a FQN for return types and parameter hints/types. if the class scanner hits a fatal for some reason, it will log the fatal and redirect you to a special page in toolbox to show you the error.

this feature is also used with the My Apps widget build option, you can turn on/off for it to analyze the app before it is downloaded/exported. it is also used in the regular download app from IPS if DT_ANALYZE is used. or you can run it manually by going to toolbox->code analyzer and running analyzer on your app.

support for containerization and wsl: I added the wsl support awhile back, but now that i'm back in linux, i'm using containers again, so i've added container support. what this means, is that if you use like the whoops editor feature (i use phpstorm, so the phpstorm protocol), it will correct the path from wsl/container to map directly to the host's path for the file (hopefully :)). it is 2023 and we should all be using some sort of containerization at this point instead of things like wamp/xamp/mamp/etc in windows/mac/linux. Docker is exellent and when used with something like devilbox makes a pretty good development environment. you can find these settings in constants->debug in toolbox.

MyApps/Dev Toybox widget: this will now show if query_log/profiler is disabled or not used. you can also hide this widget in constants->DevTools by turning on DT_HIDE_MYAPPS.