irmprovements to ClassScanner
a few changes to how the validator works. it will now check the classes that also have traits, it checks the traits against the parent classes to see if the trait is overriding methods too.  also added in a few autolints too, so methods that IPS intends to be overloaded, are ignored, where ones that are not, still get checked.