more analyzer improvements
added in a template checker, which atm only checks to see if the template will eval.

improved hook checking, to make it more inline with the class  scanner. simplified the code to make it a bit easier to understand and abstracted it out to use between the hooks and class scanner.